Clarify ghosts.md, DOC-406 (#435)
In addition to the DOC-406 request to link to an example of a ghost with nested mappings ...

I updated the links to related examples because they were permalinks to files that have moved in the latest examples repo. 
I added a link to the sum / usum feature for ghosts (and added required target / id in expr.md for cross-ref).
The grammar for ghosts was missing a production for "axioms".
Formatting fix in CLI options (DOC-294 fixup)